+++ spec/lumberjack/device/logger_file_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe Lumberjack::Device::LoggerFile do
   let(:out) { StringIO.new }
 
   it "wraps a ::Logger::LogDevice" do
-    device = Lumberjack::Device::LoggerFile.new(out, template: ":severity :message")
+    device = Lumberjack::Device::LoggerFile.new(out, template: "{{severity}} {{message}}")
     expect(device.class).to eq(Lumberjack::Device::LoggerFile)
     device.write(Lumberjack::LogEntry.new(Time.now, Logger::INFO, "Test message", nil, Process.pid, nil))
     expect(out.string.chomp).to eq("INFO Test message")
@@ -14,7 +14,7 @@ RSpec.describe Lumberjack::Device::LoggerFile do
 
   it "passes supported device options through to the underlying device" do
     expect(Logger::LogDevice).to receive(:new).with(out, shift_age: 10).and_call_original
-    Lumberjack::Device::LoggerFile.new(out, template: ":severity :message", shift_age: 10)
+    Lumberjack::Device::LoggerFile.new(out, template: "{{severity}} {{message}}", shift_age: 10)
   end
 
   it "exposes the file path for the underlying stream" do